+++ src/Config/ConfigLoader.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Phoundation\Config;
+
+interface ConfigLoader
+{
+    public function load(): array;
+}