+++ tests/TestAsset/Service/LoggerFactory.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Phoundation\Tests\TestAsset\Service;
+
+use Psr\Log\LoggerInterface;
+
+class LoggerFactory
+{
+    public function __invoke(): LoggerInterface
+    {
+        return new InMemoryLogger();
+    }
+}